Cerberus part 4 (clock graphics)
… Time is a tool you can put on the wall
Or wear it on your wrist
The past is far behind us
The future doesn't exist (oh)
… What's the time?
Its a quarter to nine, time to have a bath
What do you mean, we're already clean
Scrub, scrub, scrub 'til the water's brown
… Time is a ruler to measure the day
It doesn't go backwards, only one way
Watch it go round like a merry go round
Going so fast like a merry go round
… Let's go on a journey
A journey through all time
The time is changing all the time
It's time to go to time
… But we don't really want to
We're going to miss our show
Don't be stupid, friend
C'mon, it's time to go
… Time is old like the Victorian times
With cobbles and plague and speaking in rhymes
With cobbles and chimneys a simpler time
With cobbles and sawdust and batteries and slime
… This tree that is old has circles inside
The tree that is older has shriveled and died
The apple that is fresh is ripe to the core
And I rot over time and I'm not anymore
… Time can be told by the moon or the sun
But time flies fast when you're having fun
There's a time and a place for mucking around
Like birthdays
And camping
I'm friends with my dad
… And then what happened after the olden days?
… Time went new and got old like history
Stuff from the past went into a mystery
An old man died
But look a computer
Everything's cool
It's the future!
… Time is now, the future anew
And look at all of the wonderful things you can do
With gadgets and gizmos and email addresses
My dad is a computer
… Look at the time!
… It's quarter to eight, there's fish on my plate
It's twenty past day, there's fish on my tray
It's eleven to twelve, there's fish in the bath
It's nine thirty, there's fish everywhere, fish everywhere?
… Now you can see the importance of time
It helps us make pizza, it keeps things in line
But when did it start?
And when will it stop?
Time is important and I am a clock
If we run out of time
Then where does it go?
Is time even real?
Does anyone know?
Maybe time's just a construct of human perception
An illusion created by-
… Meh! meh! meh! meh!
Meh! meh! meh! meh!
Meh! meh! meh! meh!
Meh! meh!
… Sunrise, sunset, night and day
The changing seasons, the smell of hay
Look at your hair grow, isn't it strange?
How time makes your appearance change
… Make it stop!
It's out of my hands, I'm only a clock
Don't worry, I'm sure you'll be fine
But eventually everyone runs out of time