i found a donkey kong copypasta guys
Donkey Kong Country is truly perfect. If you do not get this amazing new generation of Donkey Kong Country madness, you are stupid. Yes, I know it's insulting, but it's also the truth. If you're a true video game fan, you will not hesitate in the slightest bit to buy this piece of gaming history.