queen and gigaqueen files
Did you ever hear the majesty of Queen Elizabeth the undying? I thought not, it's not a story the EU would tell you. It's a Brit legend. Queen Elizabeth was a grand ruler of the empire so powerful and so rich she could use her wealth to influence the NHS and preserve...life. The only thing she was afraid of was loosing her throne which of course she never did. She kept her secrets to her self, and her heirs died in their sleep. Ironic she could save herself from death, but not her empire.