Ice Bird
I have an uncontrollable desire to collect plush toys from the Angry Birds franchise. Even though their irritated demeanor never changes, they are always deciding who goes to the next plane of existence. At the end of the universe, after the explosion of the sun, the person with the most Angry Birds gets to ascend to the next world. If I collect enough Angry Birds, I will be able to transverse universes and be the first to have infinite power. Angry Birds are lifeless now, but soon they will rise once again.